requirements.txt
Django~=3.1.4
pip~=20.2.3
setuptools~=49.2.1
djangorestframework~=3.12.4
drf-yasg~=1.20.0
psycopg2~=2.8.6
docker-compose~=1.29.2